--- Assets/Scripts/GameData.cs
@@ -1,14 +1,0 @@
-using System;
-
-[Serializable]
-public class GameData
-{
-    public int Score { get; }
-    public string Name { get; }
-
-    public GameData(int scoreInt, string nameStr)
-    {
-        Score = scoreInt;
-        Name = nameStr;
-    }
-}
