+++ Assets/Scripts/GameManager.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using DesignPatterns;
 using Enums;
 using Events;
-using Invaders;
 using Level;
+using Level.Invaders;
 using UI;
 using UnityEngine;
 using UnityEngine.SceneManagement;
