--- Assets/Scripts/IKillable.cs
@@ -1,4 +1,0 @@
-public interface IKillable
-{
-        void Kill();
-}
