+++ Assets/Scripts/Level/IKillable.cs
@@ -1,0 +1,7 @@
+namespace Level
+{
+        public interface IKillable
+        {
+                void Kill();
+        }
+}
