+++ Assets/Scripts/SaveScore/GameData.cs
@@ -1,0 +1,17 @@
+using System;
+
+namespace SaveScore
+{
+    [Serializable]
+    public class GameData
+    {
+        public int Score { get; }
+        public string Name { get; }
+
+        public GameData(int scoreInt, string nameStr)
+        {
+            Score = scoreInt;
+            Name = nameStr;
+        }
+    }
+}
