+++ Assets/Scripts/UI/TextSetter.cs
@@ -1,4 +1,5 @@
 ﻿using Enums;
+using SaveScore;
 using UnityEngine;
 using UnityEngine.UI;
